feat(py): op file mv to utils
