+++ script/cities_filter.py
@@ -1,12 +1,11 @@
 import os
-import op_file
 
 current_path = os.path.dirname(__file__)
 
 cities_json_file = current_path + "/data/cities.json"
 
 # load local data
-cities_dict = op_file.load_dict_from_file(cities_json_file)
+cities_dict = utils.OperateFile.load_dict_from_file(cities_json_file)
 # print(cities_dict)
 
 
@@ -22,4 +21,4 @@ f_data = filter("JP")
 
 output = current_path + "/.augus_output"
 
-op_file.save_dict_to_file(f_data, output)
+utils.OperateFile.save_dict_to_file(f_data, output)
