+++ script/cities_filter_japan.py
@@ -1,12 +1,11 @@
 import os
-import op_file
 
 current_path = os.path.dirname(__file__)
 
 cities_json_file = current_path + "/data/gazetteer-of-japan.json"
 
 # load local data
-cities_dict = op_file.load_dict_from_file(cities_json_file)
+cities_dict = utils.OperateFile.load_dict_from_file(cities_json_file)
 # print(cities_dict)
 
 
@@ -41,4 +40,4 @@ f_data = filter()
 
 output = current_path + "/.augus_output"
 
-op_file.save_dict_to_file(f_data, output)
+utils.OperateFile.save_dict_to_file(f_data, output)
