+++ script/get_all_date.py
@@ -1,6 +1,5 @@
 import matplotlib.pyplot as plt
 import get_order_due_date
-import op_file
 
 import datetime
 from collections import Counter
@@ -29,7 +28,7 @@ while start_date < end_date:
 
     start_date += delta
 
-# op_file.write_file(res_arr_write_file)
+# utils.OperateFile.write_file(res_arr_write_file)
 
 
 # print("res_arr_for_cal:", res_arr_for_cal)
@@ -87,7 +86,7 @@ for ch in merge_all:
     str_all.append(n2str)
 
 
-# op_file.write_file(str_all)
+# utils.OperateFile.write_file(str_all)
 
 count = Counter(str_all)
