+++ script/mihoyo2.py
@@ -5,7 +5,6 @@ import os
 import requests
 from bs4 import BeautifulSoup
 import re
-import op_file
 import utils
 
 
@@ -364,4 +363,4 @@ for i in range(len(post_id)):
 
 current_path = os.path.dirname(__file__)
 filename = current_path + "/auto/mhy2.json"
-op_file.save_dict_to_file(all_info, filename)
+utils.OperateFile.save_dict_to_file(all_info, filename)
